==> 2-Data Cleaning.ipynb <==


<h1><center> Data Cleaning </center></h1>

import pandas as pd
import re
import string
import pickle
from string import punctuation
import nltk
from sklearn.feature_extraction.text import CountVectorizer
from nltk.tokenize import sent_tokenize, word_tokenize


from nltk.stem.porter import PorterStemmer
# nltk.download('wordnet')
from nltk.stem import WordNetLemmatizer
# from autocorrect import spell

## Getting Data 

## 1. Renaming (Inplace) the coloumns as per our flexibilty

df = pd.read_pickle('Data/videoInfo.pkl')
df.rename(columns=
          {"View Count": "ViewCount",
           "Comment Count": "CommentCount",
            "Uploader":"Person"
          }, inplace = True)
df.head()

Extracting only person and comments he or she received from each of the videos:

## 2. nine coloums or features for 16 videos

df.shape

df.describe

## Here we have comments as per in individual video

person_comments = df[['Person','Comments']]
person_comments.head()

Some videos I chose are uploaded by uploader, but subjects are different in the content. Regardless, I would like to analyze comments in the videos. 
Hence, I am replacing the person name by subject of the video.

## 3. Changing the name of some subjects as uploader are different in this case but the videos are of our concern topic and of our concering celebrites also

## Checkpint ----

#Replacing Person with actual subject in the video
person_comments = person_comments.replace({'Person' : {'Charlie Ayee': 'Steven Assanti', 
                               'AlienRadioStation' : 'Steven Assanti',
                               'rebecca' : 'Rebecca Black',
                               'nowthisisliving':'Shannon Beveridge',
                               'RoughChop':'Richard Gale'}})
person_comments['Person'].to_frame().T

**Combining subjects in one row**:

Each subject has multiple videos and comment pairs. I am interested in finding what kind of comments each uploader gets. In this light I can combine comments from multiple videos into one for a user.

## 4. Combining all videos comment of a particular uploader into one for determing the type of comment each uploader gets

person_comments_grouped=person_comments.groupby(['Person'])['Comments'].apply(','.join).reset_index()
person_comments_grouped

#Pickling grouped unclean data for later use
person_comments_grouped.to_pickle("person_comments_grouped.pkl")

# Looking at a sample text:
person_comments_grouped['Comments'][1][:1000]

## Cleaning Data

**Common data cleaning steps on all text:**

* Make text lower case
* Remove punctuation
* Remove numerical values
* Remove common non-sensical text
* Tokenize text
* Remove stop words
* Removing emojis and icons

**More data cleaning steps after tokenization:**

* Stemming Lemmatization
* Deal with typos

# Text Cleaning

def textCleaning(text):
    text = re.sub('[%s]' % re.escape(punctuation), ' ', text)  #remove punctuation
    text = re.sub('\w*\d\w*', ' ', text)                       #remove words with numbers
    text = re.sub('[‘’“”…]', ' ', text)
    text = text.lower()                                        # make lowercase
    text = re.sub('\n', ' ', text)                             # remove new line
    return text


text_cleaner = lambda x: textCleaning(x)


## Checkpoint ______

#Removing Emojis

def removing_emojis(text):
    # Emojis pattern
    emoji_pattern = re.compile("["
                    u"\U0001F600-\U0001F64F"  # emoticons
                    u"\U0001F300-\U0001F5FF"  # symbols & pictographs
                    u"\U0001F680-\U0001F6FF"  # transport & map symbols
                    u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                    u"\U00002702-\U000027B0"
                    u"\U000024C2-\U0001F251"
                    u"\U0001f926-\U0001f937"
                    u'\U00010000-\U0010ffff'
                    u"\u200d"
                    u"\u2640-\u2642"
                    u"\u2600-\u2B55"
                    u"\u23cf"
                    u"\u23e9"
                    u"\u231a"
                    u"\u3030"
                    u"\ufe0f"
        "]+", flags=re.UNICODE)
    return emoji_pattern.sub(r' ', text) 

emojiRemoval = lambda x: removing_emojis(x)

#Lemmatization
lemma = WordNetLemmatizer()
def lemmatization(text):
    words = word_tokenize(text)
    lemmatized_text=[]
    for w in words:
        w= lemma.lemmatize(w)
        lemmatized_text.append(w)
    return ' '.join(lemmatized_text)

lemmatizor = lambda x: lemmatization(x)


#Stemming
porter = PorterStemmer()
def stemming(text):
    words = word_tokenize(text)
    stemmed_text = []
    for w in words:
        w = porter.stem(w)
        stemmed_text.append(w)
    return ' '.join(stemmed_text)

stemmor = lambda x: stemming(x)


#Spell Checking
def spellCheck(text):
    autocorrected_text = spell(text)
    return autocorrected_text

spell_checker = lambda x: spellCheck(x)

## Organizing Data 


### Corpus 

Corpus is a collection of text. Here we are creating collection of cleaned texts. 

Let us apply different text cleaning functions we created above.

For now, I am skipping lemmatization, stemming and spell checking. The reason for this is because they tend to correct per change spellings of texts. In this project, we are looking for abuse words too. We don't want such words to be changed.

## 5. Skipping the stemming and lemmitization part as here we have abuse words too , and we dont want there spelling to get changed while cleaning the text

comment_clean= person_comments_grouped['Comments'].apply(text_cleaner).apply(emojiRemoval)
               #.apply(lemmatizor).apply(stemmor).apply(spell_checker)

person_comments_cleaned= person_comments_grouped.drop(['Comments'], axis=1)
person_comments_cleaned.insert(1, 'Cleaned_Comments', comment_clean)

person_comments_cleaned

## 6. We can see that comments got cleaned but the spelling is not changed as we have not done stemming and lemmitization

# Looking at a sample text:
person_comments_cleaned['Cleaned_Comments'][1][:1000]

#Pickling cleaned data for later use
person_comments_cleaned.to_pickle("person_comments_cleaned.pkl")

## 7. Zipping the person and cleaned comments into a dictonary and then combing them all into a single large piece of text as we have cleaned the text above so theri might be white space in between them then again converting it back to dataframe for the further analysic

#  Converting to the dictonary
personComments_dict = dict(zip(person_comments_cleaned.Person, person_comments_cleaned.Cleaned_Comments))


# This function converts strings into single large peice of text removing white spaces
def combine_text(list_of_text):
    combined_text = ''.join(list_of_text)
    return combined_text



person_comments_cleaned_kvp = {key: [combine_text(value)] for (key, value) in personComments_dict.items()}

# Again converting back to the data frame
personComments_corpus = pd.DataFrame.from_dict(person_comments_cleaned_kvp).transpose()

personComments_corpus.columns = ['comments']
# Sort on the basis of index values
personComments_corpus = personComments_corpus.sort_index()
personComments_corpus

## 7.Pickling the above file into local system for later use

#Pickling corpus for later use
personComments_corpus.to_pickle('personComments_corpus.pkl')

## checkpoint

# Major goal of this part
## 8. Finally making a  Document Term Matrix (DTM)

Document Term Matrix is word counts in matrix format

# Creating document-term matrix using CountVectorizer and excluding common English stop words

cv = CountVectorizer(stop_words='english')
data_cv = cv.fit_transform(personComments_corpus.comments)
personComments_dtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
personComments_dtm.index = personComments_corpus.index
personComments_dtm

personComments_dtm.to_pickle("personComments_dtm.pkl")
pickle.dump(cv, open("cv.pkl", "wb"))
