==> 4- Topic Modeling.ipynb <==


<h1><center> Topic Modeling using LDA</center></h1>

import pandas as pd
import pickle

from gensim import matutils, models
import scipy.sparse
import nltk
#nltk.download('averaged_perceptron_tagger')
from nltk import word_tokenize, pos_tag

from sklearn.feature_extraction import text
from sklearn.feature_extraction.text import CountVectorizer

Topic modeling is looking at words in a text and determining what topic is the text all about. Here we will try to find topics using LDA (Latent Dirichlet Allocation).  We will go through several iterations in determining topics. For now, I will use parts of speech in choosing words in texts for the iterations.

### Iteration 1: All Text

#Term Document Matrix (tdm) -Gensim needs transpose of dtm
personComments_topic1 = pd.read_pickle('personComments_dtm2.pkl').transpose()
personComments_topic1.head()

# Putting tdm into gensim format, 
sparse_counts = scipy.sparse.csr_matrix(personComments_topic1)
corpus = matutils.Sparse2Corpus(sparse_counts)

# Dictionary of the all terms and their respective location in the tdm
cv = pickle.load(open("cv_stop.pkl", "rb"))
id2word = dict((v, k) for k, v in cv.vocabulary_.items())

**LDA** (Latent Dirichlet Allocation)

LDA is an example of topic model which builds a topic per document model and words per topic model.

# Specifying LDA with the number of topics and the number of passes
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=20)
lda.print_topics()

### Iteration 2: Nouns

#Function to return nouns from text
def nouns(text):
    '''Given a string of text, tokenize the text and pull out only the nouns.'''
    is_noun = lambda pos: pos[:2] == 'NN'
    tokenized = word_tokenize(text)
    all_nouns = [word for (word, pos) in pos_tag(tokenized) if is_noun(pos)] 
    return ' '.join(all_nouns)

#Reading Clean data
personComments_corpus = pd.read_pickle('personComments_corpus.pkl')
personComments_corpus

# Filtering noun from clean data
personComments_nouns = pd.DataFrame(personComments_corpus.comments.apply(nouns) )
personComments_nouns

# Re-add the additional stop words since we are recreating the document-term matrix
add_stop_words = ['like', 'just', 'don', 'br', 'video', 'know', 'people', 'love', 'really', 'dani', 
 'danielle', 'madison', 'rebecca', 'shannon', 'really', 'richard', 'casey', 'emily', 'steven']
stop_words = text.ENGLISH_STOP_WORDS.union(add_stop_words)




# Recreating Document-Term matrix with only nouns

cv_noun = CountVectorizer(stop_words=stop_words)
data_cv_noun = cv_noun.fit_transform(personComments_nouns.comments)
personComments_dtm_nouns= pd.DataFrame(data_cv_noun.toarray(), columns=cv_noun.get_feature_names())
personComments_dtm_nouns.index = personComments_nouns.index
personComments_dtm_nouns

# Creating the gensim corpus with only noun
corpus_noun = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(personComments_dtm_nouns.transpose()))

# Create the vocabulary dictionary
id2word_noun = dict((v, k) for k, v in cv_noun.vocabulary_.items())

# Creating LDA with nouns with 4 topics
lda_noun = models.LdaModel(corpus=corpus_noun, num_topics=4, id2word=id2word_noun, passes=20)
lda_noun.print_topics()

### Iteration 3: Nouns and Adjectives

#Function to return nouns and adjectives from text
def nouns_adj(text):
    '''Given a string of text, tokenize the text and pull out only the nouns and adjectives.'''
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    tokenized = word_tokenize(text)
    nouns_adj = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)] 
    return ' '.join(nouns_adj)

# Applying nouns_adj functions to filter nouns and adjective
personComments_nounsAdj = pd.DataFrame(personComments_corpus.comments.apply(nouns_adj))
personComments_nounsAdj

# Create a new document-term matrix using only nouns and adjectives, also remove common words with max_df
cv_nouns_adj = CountVectorizer(stop_words=stop_words, max_df=.8)
data_cv_nouns_adj = cv_nouns_adj.fit_transform(personComments_nounsAdj.comments)
personComments_dtm_nounsAdj = pd.DataFrame(data_cv_nouns_adj.toarray(), columns=cv_nouns_adj.get_feature_names())
personComments_dtm_nounsAdj.index = personComments_nounsAdj.index
personComments_dtm_nounsAdj

# Create the gensim corpus
corpus_nouns_adj = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(personComments_dtm_nounsAdj.transpose()))

# Create the vocabulary dictionary
id2word_nouns_adj = dict((v, k) for k, v in cv_nouns_adj.vocabulary_.items())

# LDA topic modeling with 4 topics
lda_nouns_adj = models.LdaModel(corpus=corpus_nouns_adj, num_topics=4, id2word=id2word_nouns_adj, passes=80)
lda_nouns_adj.print_topics()

# Let's take a look at which topics comments for each uploader contains
corpus_transformed = lda_nouns_adj[corpus_nouns_adj]
list(zip([a for [(a,b)] in corpus_transformed], personComments_dtm_nounsAdj.index))

### Topics and People


   
* **Topic 0:** song/music/voice, Friday/party/weekend

  Rebecca Black   
    
    
* **Topic 1:** beautiful/perfect/amazing song/music/voice 

  Madison Beer


* **Topic 2:** gay/lesbian, sex, kid, fat, victim, b**ch

  Shannon Beveridge, Steven Assanti, Richard Gale
 
 
 * **Topic 3:** arab, gold, girls, family, beautiful, pregnant

    Emily Ann Shaheen, Madeline & Eric


#### Insights

From topics people talk about in the comments for each of the people , we can see that yet **Shannon Beveridge, Steven Assanti, Richard Gale** are prone to harsh comments than any other people in the list.
